SC2_Co-op_Resource/requirements.txt
beautifulsoup4>=4.12.2
selenium>=4.15.2
requests>=2.31.0
Pillow>=10.1.0 